Update class comment on EventDataBatch class
Adjusting top comment, as SendBatch and SendBatchAsync methods don't exist on EventHubClient, the methods are just called Send and SendAsync and take either EventData or IEnumerable<EventData>.
In addition the comment makes it not clear why a dev should use this class. Why not just create an IEnumerable<EventData> on your own? A dev should use this class because it takes care of the size limit, that's the main reason and it should be in the header.