Update EventDataBatch.cs
Adjusting top comment, as SendBatch and SendBatchAsync methods don't exist on EventHubClient anymore, the methods are called Send and SendAsync and take either a single EventData or an IEnumerable<EventData>.
In addition the comment makes it not clear why a dev should use this class. Why not just create an IEnumerable<EventData> on your own? A dev should use this class because it takes care of the size limit, that's the main reason and it should be in the header.